+++ K8sUtils/private/Write-PodEvent.ps1
@@ -57,7 +57,7 @@ function Write-PodEvent {
     if ($errors -and $FilterStartupWarnings) {
         $errors = $errors | Where-Object { $_ -notlike "Startup probe failed:*" }
     }
-    $events | Select-Object type, reason, message | Write-Plain
+    $events | Select-Object type, reason, message | Out-String | Write-Plain
     Write-Footer "End events for $Prefix $PodName" -LogLevel ($errors ? "error" : $LogLevel)
     if ($PassThru) {
         return $errors